AI-Python/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI Providers
openai==1.3.0
groq==0.4.1
google-generativeai==0.3.0

# Audio/Video Processing
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1
python-magic==0.4.27

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
